requirements.txt
python-decouple==3.6
vk_maria==3.0.7
requests~=2.28.2